Fix problem loading gem caused by name change
The gem name was changed to activerecord-find_only, to match the activerecord gem name.
But all of the files are in the `active_record` directory, to match the
ruby casing convention. The bundler auto-require was looking for
activerecord/find_only.rb to load and wasn't finding anything. This adds
a redirect file so it can be required as `activerecord/find_only` or
`active_record/find_only`.